src/requirements.txt
antlr4-python3-runtime==4.13.0
antlr4-tools==0.2
install-jdk==1.0.4
networkx==3.1
